+++ packages/frontend-logic/src/shared/services/fixedLoc.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { NewFixedLocBody, EditFixedLocBody, FixedLocDT, EditManyFixedLocBody } from '@m-cafe-app/utils';
-import { RequestOptions } from '../types/requestOptions';
+import { RequestOptions } from '../../types/requestOptions';
 
 /**
  * Should not ever be used. FixedLocs are fixed, only updatable. Delete this?
